Adding one extra rewrite for blog favicon. (#65)
I've made it to go to the same as the web, since it's the same domain at
this stage. I am aware of the fact that I'm rewriting .png into .ico
file, but all the correct headers will come through therefore browser
will apply it regardless of the file extension.